⚡️ make navbar server component and add navItem client
