+++ app/components/navbar/navbar.tsx
@@ -1,7 +1,11 @@
 // components/Navbar.tsx
 import { getDetails } from "@/app/types/getDetails";
+import { Button } from "@/components/ui/button";
+import { CircleUser, Github, Home, LibraryBig, Tv } from "lucide-react";
+import { ModeToggle } from "../modeToggle";
+import SearchBar from "../searchbar";
 import { MobileNavbarComponent } from "./mobilenavbar";
-import NavbarClient from "./navbarClient";
+import NavItem from "./navItem";
 async function fetchAllItems() {
 	try {
 		const itemsDetails = await getDetails();
@@ -21,13 +25,50 @@ async function fetchAllItems() {
 		return [];
 	}
 }
+const language = process.env.DEFAULT_LANGUAGE;
+const data = require(`@/locales/${language}.json`);
 
 export default async function Navbar() {
 	const details = await fetchAllItems();
 	return (
 		<nav className="bg-white/25 dark:bg-black/25 backdrop-filter backdrop-blur-lg p-2 shadow-md md:block hidden">
 			<div className="flex justify-between items-center max-w-7xl mx-auto">
-				<NavbarClient details={details} />
+				<>
+					<div className="flex items-center">
+						<NavItem href="/">
+							<Home size={20} />
+							<span className="ml-2">{data.navbar.home}</span>
+						</NavItem>
+						<NavItem href="/profil">
+							<CircleUser size={20} />
+							<span className="ml-2">{data.navbar.profil}</span>
+						</NavItem>
+						<NavItem href="/live">
+							<Tv size={20} />
+							<span className="ml-2">Live</span>
+						</NavItem>
+					</div>
+					<div className="flex items-center space-x-4">
+						<NavItem href="/search">
+							<LibraryBig size={20} />
+							<span className="ml-2"> {data.navbar.catalog}</span>
+						</NavItem>
+						<SearchBar details={details} />
+						<a
+							href="https://github.com/FlorianDevv/MangaRead"
+							target="_blank"
+							rel="noopener noreferrer"
+							aria-label="View source on GitHub"
+						>
+							<Button variant="ghost">
+								<p className="transition-colors hover:text-foreground/80">
+									<Github size={20} />
+								</p>
+							</Button>
+						</a>
+						<ModeToggle />
+					</div>
+				</>
 			</div>
 		</nav>
 	);
